Split CSS into separate files, update references to them in templates
Rather than 'content' and 'home' there is now 'main', 'home' and 'news'.
The 'main.css' file contains all of the CSS to produce the layout and
style the content of a basic page. 'home.css' modifies some of this
styling for the home page and 'news.css' does this for the news page.