Added caching of parsed $content objects.
When a content page is requested, it is checked to see if the client's
cache is valid, and if so, send a 304. If not, it checks to see if the
corresponding Content object is in the server's cache (memcached) and if
so, serve that. If not, parse the file, add it to memcache and serve it.