Style 'p' tags inside info/warning boxes
Allows markdown to be used inside the info/warning boxes. Adding the
'markdown="1"' attribute to a div will make markdown wrap the text in 'p' tags
which messed up the styling. It now looks the same as it did without the
markdown attribute but you can use all of markdown's features.