Made search.php spit out an error message if no query is given.
Previously, the behaviour was to redirect back to $_SERVER['HTTP_REFERER'].
There were two problems associated with this: one was the obvious violation
of the principle of least surprise - one would not normally expect to be
simply redirected back in case of an error. The other was the use of
$_SERVER['HTTP_REFERER'] without first checking that it exists.

The behaviour now is to spit out a 400 Bad Request, and a simple plain-
text message informing the user of the problem.

Signed-off-by: Chris Kirkham <chrisjameskirkham@gmail.com>
Signed-off-by: Alastair Lynn <arplynn@gmail.com>