Fix what I suspect to be a rendering bug in Firefox
A paragraph with italic text in it overflows the block on the right hand
side. As the paragraph I noticed this problem on had overflow set to hidden
some of the letters were being cut off. Adding a small bit of padding to the
right of the paragraph fixes the problem.