Add a note box to the yield and coroutines page to prevent confusion
The page uses 'yield 1' a lot to describe how generators work in a simple way.
In the past this would have acted as a time delay on the robot but the new
syntax for that is different. The note is just to point out to people reading
the docs that 'yield 1' isn't valid.