Updated the URL for the firmware source on the servo board.
The link was pointing to a 'pwm-fw' repository, which was undergoing a
critical failure to exist due to its being renamed to 'servo-fw' in line
with the name change of the board.

This commit fixes the link to point at the new repository on the
docs/kit/servo_board page.

Note this commit affects only the English version of the content - the French
content may need changing to reflect this.

Signed-off-by: Alastair Lynn <arplynn@gmail.com>