Fix bug in munging of links
Links in markdown are munged to add in the root path so that the site
works when cloned into a non-root dir of a web server. The replace only
matched the first link on a line and the rest of the links on a line were
un-modified causing them to link to the root of the web server.