Add a parser for PHP
It is now possible to specifiy 'php' as a CONTENT_TYPE in a page file.
This runs the main contents of the file through 'eval' and captures
the output of any php executed. The file must contain html with inline
php.

The follwing is an example file using the php CONTENT_TYPE:

//TITLE:        Bees
//CONTENT_TYPE: php

<h1>#Bees in my Eyes#</h1>
<?php echo "I have bees in my eyes." ?>