Isolate the display of Team Status from the underlying object format.
Add a common method that constructs the information about a team, and
a helper class for wrapping the new backend format. Modify the template
to use empty() which is a better check than != "", and allows me to use
null for things that aren't present, which is preferable.

Make the overview page actually show a thumbnail of the team's image,
rather than some ponies. The IDE can trivially also create the thumbnail
when it resizes the original.

Add a sad robot thumbnail for use when teams don't have an image yet.