Make the team status images work.
Make the info laoder look in the right place to detect whether or not
there should be an image to display.

Also re-arrange how the root of the images paths is determined.
Unfortunately, it gets used in two places that have slightly different
rewriting rules so we need to add the root manually in the template
but not include it in the defined root.