Switch to accepting old-style team numbers as well as team ids.
We're not using the new style ids yet, and this was causing issues
when interacting with numeric ids. Also make the prefixing .* less
greedy by adding a ? so that it doesn't swallow most of the id too
(this is only needed now we're less specific about the makeup of the
ids).