Create a LUT for college names/URLs from team IDs
This allows the team info pages to show the college name and link to the
website of the college a team is from. Even though this data is sort of
available in LDAP (teams aren't explicitly linked to colleges) I've decided
to partially duplicate this information to save adding LDAP access to srweb.