Populate the team-college LUT with data from trac.
This information is lifted directly from the 2012/Schools trac page:
https://www.studentrobotics.org/trac/wiki/2012/Schools. Any corrections
should therefore probably be applied in both places.

Note that this is currently assuming that the LUT needs to use the
team's ids rather than numbers. In the event that this is wrong,
the teams are in the same order as the trac page, and thus also in
numeric order.