Don't serve 304 Not Modified for php type pages
As these are dynamically generated there's no way of knowing if the
generated content will be different to what is cached in the browser so
we should always send the page.

Ideally we wouldn't cache the contents on the server side either but
this is more a more complex issue. At the moment the caching on the
server tries not to read the file if at all possible, to find out if the
page is dynamic means it'd always have to read the file, even if there's
a cached copy. Probably should split the reading/parsing into two steps.