Add a bunch of fixes for older browsers.
This changes the search button to being an image, rather than an
image styled button, and adds a div around the main list of links
(which is then floated right instead of the ul directly).
Also ensure that the headings on the sponsors page clear the images
above them rather than being shunted to the side.