Fix #1133: Prevent UTF-8 BOMs causing rendering issues.
This is a very simplistic fix, and does nothing other than remove
the BOM from the line. It assumes that the file is valid, and that
the rest of srweb can cope with the file containing Unicode characters.

This fix intentionally doesn't do anything about BOMs for other encoding
schemes, since they would likely require conversion and/or other special
hanlding. As and when we decide to implement support for them, this should
probably be removed in favour for more generalised handling.