Change news page to use standard bottom links
As part of the development of the team-status stuff I created a bunch of
CSS classes for top/bottom tab-style links that are useful for
previous/next or back buttons. This style of link is now used on the
team-status pages and the news article pages. This commit alters the
'older news'/'newer news' links on the news page to match the rest of
the site.