Make news item titles links to the article pages
Each news item has its own article page but it's currently quite hard to
find how to get to it. You either have to click through to it from the
homepage or notice, and understand the grey 'permalink' link underneath
the news item on the news page. This makes the news item title into a
link to the article page so that the website is a bit more intuative to
use.