Revert "Add /voting to non-srweb-managed urls."
This reverts commit 127d816a21fb427dc4a43353ad74c9244285218a.

The auto git updating stuff works; however it turns out that the voting
scripts depend on suexec, which is fine. But suexec isn't configurable
enough to enable specific users for specific scripts, so it looks like
we can't dump the voting scripts into srweb, and it'll have to go in a
userdir instead.