Fix #1407: Add support for a local config file.
The file is called local.config.inc.php, and can override anything
that's define()d in config.inc.php. I don't think there's a need to
be able to overwrite the variables defined as well (at least for the
moment).

By changing the define()s in config.inc.php to check for the existence
of the constant before trying to override it, we avoid a potential
bunch of Notices from php.