Disable the signup form & add a constant that can enable it.
This hides the signup form for this year, while adding a simple way
that it can be re-enabled. I've made the default be the 'watch this
space for details' since the sign-up process will likely change a
bit each year. That said, the underway message will likely want
tweaking each year with the year number & article too.